Add files via upload
A/B tests are very commonly performed by data analysts and data scientists. For this project, we will be working  to understand the results of an A/B test run by an e-commerce website our goal is to work through this notebook to help the company understand if they should:
Implement the new webpage,
Keep the old webpage, or
Perhaps run the experiment longer to make their decision.